+++ datanode/gateway/graphql/mocks/mocks.go
@@ -820,6 +820,26 @@ func (mr *MockTradingDataServiceClientV2MockRecorder) GetParty(arg0, arg1 interf
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetParty", reflect.TypeOf((*MockTradingDataServiceClientV2)(nil).GetParty), varargs...)
 }
 
+// GetPartyActivityStreak mocks base method.
+func (m *MockTradingDataServiceClientV2) GetPartyActivityStreak(arg0 context.Context, arg1 *v2.GetPartyActivityStreakRequest, arg2 ...grpc.CallOption) (*v2.GetPartyActivityStreakResponse, error) {
+	m.ctrl.T.Helper()
+	varargs := []interface{}{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "GetPartyActivityStreak", varargs...)
+	ret0, _ := ret[0].(*v2.GetPartyActivityStreakResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetPartyActivityStreak indicates an expected call of GetPartyActivityStreak.
+func (mr *MockTradingDataServiceClientV2MockRecorder) GetPartyActivityStreak(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetPartyActivityStreak", reflect.TypeOf((*MockTradingDataServiceClientV2)(nil).GetPartyActivityStreak), varargs...)
+}
+
 // GetProtocolUpgradeStatus mocks base method.
 func (m *MockTradingDataServiceClientV2) GetProtocolUpgradeStatus(arg0 context.Context, arg1 *v2.GetProtocolUpgradeStatusRequest, arg2 ...grpc.CallOption) (*v2.GetProtocolUpgradeStatusResponse, error) {
 	m.ctrl.T.Helper()